requirements.txt
certifi==2020.12.5
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.0
Django==3.1.5
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
idna==2.10
pycparser==2.20
PyJWT==2.0.1
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
